Fix create and delete funcs of namesapces
