+++ pkg/utils/common/controller.go
@@ -19,6 +19,7 @@ import (
 	"github.com/redhat-appstudio/e2e-tests/pkg/utils"
 	appsv1 "k8s.io/api/apps/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
+	k8sErrors "k8s.io/apimachinery/pkg/api/errors"
 	rclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -226,4 +227,52 @@ func (s *SuiteController) GetConfigMap(name, namespace string) (*corev1.ConfigMa
 
 func (s *SuiteController) DeleteConfigMap(name, namespace string) error {
 	return s.KubeInterface().CoreV1().ConfigMaps(namespace).Delete(context.TODO(), name, metav1.DeleteOptions{})
+}
+
+// DeleteNamespace deletes the give namespace.
+func (s *SuiteController) DeleteNamespace(namespace string) error {
+	_, err := s.KubeInterface().CoreV1().Namespaces().Get(context.TODO(), namespace, metav1.GetOptions{})
+
+	if err != nil && !k8sErrors.IsNotFound(err) {
+		return fmt.Errorf("could not check for namespace existence")
+	}
+
+	return s.KubeInterface().CoreV1().Namespaces().Delete(context.TODO(), namespace, metav1.DeleteOptions{})
+}
+
+// CreateTestNamespace creates a namespace where Application and Component CR will be created
+func (h *SuiteController) CreateTestNamespace(name string) (*corev1.Namespace, error) {
+
+	// Check if the E2E test namespace already exists
+	ns, err := h.KubeInterface().CoreV1().Namespaces().Get(context.TODO(), name, metav1.GetOptions{})
+
+	if err != nil {
+		if k8sErrors.IsNotFound(err) {
+			// Create the E2E test namespace if it doesn't exist
+			nsTemplate := corev1.Namespace{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:   name,
+					Labels: map[string]string{constants.ArgoCDLabelKey: constants.ArgoCDLabelValue},
+				}}
+			ns, err = h.KubeInterface().CoreV1().Namespaces().Create(context.TODO(), &nsTemplate, metav1.CreateOptions{})
+			if err != nil {
+				return nil, fmt.Errorf("error when creating %s namespace: %v", name, err)
+			}
+		} else {
+			return nil, fmt.Errorf("error when getting the '%s' namespace: %v", name, err)
+		}
+	} else {
+		// Check whether the test namespace contains correct label
+		if val, ok := ns.Labels[constants.ArgoCDLabelKey]; ok && val == constants.ArgoCDLabelValue {
+			return ns, nil
+		}
+		// Update test namespace labels in case they are missing argoCD label
+		ns.Labels[constants.ArgoCDLabelKey] = constants.ArgoCDLabelValue
+		ns, err = h.KubeInterface().CoreV1().Namespaces().Update(context.TODO(), ns, metav1.UpdateOptions{})
+		if err != nil {
+			return nil, fmt.Errorf("error when updating labels in '%s' namespace: %v", name, err)
+		}
+	}
+
+	return ns, nil
 }
